CustomEvent elemento utils and applied as weird-action-bus
The idea is a pretty poor and badly decoupled model based on redux. Buy yep, it is still just a weird strategy, anyway it serves as a demonstration on CustomEvent usage. Also applied RxWidget to ApplicationWidget, although this widgets is never detached so not as useful as it might be.